fix/harden-web-extractor-and-frame-types
Address PR review:
- WebServerLogExtractor: don't comma-split the request URI; wrap saveAll in
  try/catch (graceful degradation, matching DNS); free finished streams from
  pendingByStream.
- PcapParserService: read frame.number from the last column (robust to a '|' in
  an earlier field like Info).
- buildTlsInfo: prefer dst-side (server) conversations but fall back to all of
  the host's conversations when the flow was captured reversed.
- Frame columns widened INTEGER->BIGINT and Integer->Long end-to-end to match
  packet_number.
- ConversationPage: validate ?packet= as a positive safe integer.

+++ backend/src/main/java/com/tracepcap/analysis/entity/DnsQueryLogEntity.java
@@ -63,5 +63,5 @@ public class DnsQueryLogEntity {
 
   /** frame.number of the first response packet for this query — links the row to its packet. */
   @Column(name = "sample_frame")
-  private Integer sampleFrame;
+  private Long sampleFrame;
 }
